+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+- Message passed in update cart product when editing subscription rather than throwing an error
+
 ## [1.2.1] - 2019-01-28
 ### Added
 - Portal around the text for Recharge AddToCartBar Component